added tests

+++ utilities/helper.py
@@ -5,17 +5,22 @@ from datetime import timedelta
 
 
 def get_data(symbol, timestamp, year, month):
-    # Get the absolute path of the current script
+    # Get the absolute path of the parent directory of the current script
     script_directory = os.path.dirname(os.path.abspath(__file__))
-    data_directory = os.path.join(script_directory, "data")
+    parent_directory = os.path.dirname(script_directory)
+
+    # Construct the path to the 'data' directory
+    data_directory = os.path.join(parent_directory, "data")
     timestamp_directory = os.path.join(data_directory, timestamp)
+
     # Create the file path
     csv_filename = f"{symbol}-{timestamp}-{year}-{month}.csv"
     csv_path = os.path.join(timestamp_directory, csv_filename)
+
     return csv_path
 
 
-def get_zipl_files_path(symbol, month, timestamp):
+def get_zip_files_path(symbol, month, timestamp):
     # Get the absolute path of the current script
     script_directory = os.path.dirname(os.path.abspath(__file__))
     data_directory = os.path.join(script_directory, "zip_files")
@@ -26,13 +31,19 @@ def get_zipl_files_path(symbol, month, timestamp):
 
 
 def move_downloaded_file_to_required_folder(file_path: str, symbol):
-    # Get the absolute path of the current script
+    # Get the absolute path of the current script (located under 'utilities')
     script_directory = os.path.dirname(os.path.abspath(__file__))
-    # Get the current path of the file
-    csv_file_current_path = os.path.join(script_directory, file_path)
-    # Define the data directory
-    data_directory = os.path.join(script_directory, "data")
-    # Define the timestamp directory
+
+    # Define the 'data' directory which is assumed to be in the parent directory of 'utilities'
+    parent_directory = os.path.dirname(script_directory)
+
+    # Get the current path of the file (assuming 'file_path' is relative to 'utilities')
+    csv_file_current_path = os.path.join(parent_directory, file_path)
+
+    # get data folder
+    data_directory = os.path.join(parent_directory, "data")
+
+    # Define the timestamp directory within 'data'
     timestamp_directory = os.path.join(data_directory, file_path.split("-")[1])
 
     # Ensure the timestamp directory exists
@@ -43,9 +54,11 @@ def move_downloaded_file_to_required_folder(file_path: str, symbol):
     required_path = os.path.join(timestamp_directory, csv_filename)
 
     # Move the file to the required folder
-    shutil.move(csv_file_current_path, required_path)
-
-    print(f"File moved to: {required_path}")
+    if os.path.exists(csv_file_current_path):
+        shutil.move(csv_file_current_path, required_path)
+        print(f"File moved to: {required_path}")
+    else:
+        print(f"Source file does not exist: {csv_file_current_path}")
 
 
 def clean_and_save_data(file_path: str, start_date: str, end_date: str):
